requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiohttp-apispec==2.2.3
aiohttp-cors==0.7.0
aiosignal==1.3.1
aiosmtplib==1.1.7
aiosqlite==0.20.0
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.6.2
apispec==3.3.2
appdirs==1.4.4
APScheduler==3.10.4
aries-askar==0.3.0
aries_cloudagent==0.11.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.26.0
attrs==22.1.0
Authlib==1.1.0
base58==2.1.1
bcrypt==4.0.1
beautifulsoup4==4.12.0
billiard==4.2.0
bitarray==2.9.2
blinker==1.5
boto3==1.34.84
botocore==1.34.84
cached-property==1.5.2
cachetools==5.2.0
celery==5.3.6
certifi==2022.9.24
cffi==1.15.1
channels==4.1.0
chardet==3.0.4
charset-normalizer==2.1.1
cheroot==10.0.0
ckzg==1.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coinbase-commerce==1.0.1
colorama==0.4.6
ConfigArgParse==1.5.5
CredSLayer==0.1.3
cryptography==38.0.3
cssbeautifier==1.15.1
cytoolz==0.12.2
decorator==5.1.1
deepmerge==0.3.0
defusedxml==0.7.1
Deprecated==1.2.14
did-peer-2==0.1.2
dj-database-url==1.2.0
Django==4.0.1
django-allauth==0.53.1
django-bootstrap4==22.3
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-environ==0.11.2
django-jazzmin==2.6.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.4.0
django-smtp-ssl==1.0
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
djlint==1.34.1
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.21.7
dynaconf==3.2.5
ecdsa==0.18.0
EditorConfig==0.12.4
email-validator==1.3.0
environ==1.0
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.1.0
eth-utils==4.1.0
eth_abi==5.1.0
exceptiongroup==1.0.4
fastapi==0.87.0
fastapi-mail==1.2.1
fastapi-pagination==0.11.3
fastapi-utils==0.2.1
flower==2.0.1
frozendict==2.3.10
frozenlist==1.4.1
google-auth==2.14.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
h2==3.2.0
hexbytes==0.3.1
honcho==1.1.0
hpack==3.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httpcore==0.16.1
httpx==0.23.1
humanize==4.9.0
hvac==2.1.0
hyperframe==5.2.0
idna==3.4
indy-credx==1.1.1
indy-vdr==0.4.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
jaraco.functools==4.0.0
Jinja2==3.1.2
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.22
jsonpath-ng==1.5.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.0
knox==0.1.14
kombu==5.3.4
loguru==0.7.2
lru-dict==1.2.0
lxml==4.9.3
Mako==1.2.4
Markdown==3.1.1
MarkupSafe==2.1.1
marshmallow==3.20.1
more-itertools==10.2.0
multidict==6.0.5
mysql-connector-python==8.0.31
nest-asyncio==1.5.8
oauthlib==3.2.2
orjson==3.10.1
packaging==23.2
parsimonious==0.10.0
passlib==1.7.4
pathspec==0.12.1
paystack==1.5.0
Pillow==9.3.0
pluggy==1.0.0
ply==3.11
portalocker==2.7.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.2
pydantic_core==2.16.3
pydid==0.4.2
Pygments==2.17.2
PyJWT==2.8.0
PyLD==2.0.3
pymongo==3.12.3
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.0.9
pyshark==0.6
pytest==7.2.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.5
python3-indy==1.16.0.post286
python3-openid==3.2.0
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
qrcode==6.1
redis==5.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rlp==4.0.0
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
schedule==1.2.1
sd-jwt==0.10.3
sentry-sdk==1.11.1
shortuuid==1.0.13
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.44
sqlparse==0.4.4
sse-starlette==1.2.1
starlette==0.21.0
stripe==5.0.0
termcolor==2.3.0
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tqdm==4.66.2
typing_extensions==4.4.0
tzdata==2024.1
tzlocal==5.2
unflatten==0.1.1
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.19.0
validators==0.28.0
vine==5.1.0
wcwidth==0.2.13
web.py==0.62
web3==6.17.0
webargs==5.5.3
websockets==12.0
whitenoise==6.2.0
wrapt==1.16.0
yarl==1.9.4
paystackapi